otel as a plugin
First draft of what this might look like.

fixes: #399
