+++ apollo-router/src/configuration/testdata/config_opentelemetry_jaeger_basic.yml
@@ -3,5 +3,3 @@ server:
 subgraphs:
   example:
     routing_url: http://foo
-opentelemetry:
-  jaeger:
