+++ apollo-router/src/configuration/testdata/config_opentelemetry_otlp_tracing_grpc_common.yml
@@ -3,11 +3,3 @@ server:
 subgraphs:
   example:
     routing_url: http://foo
-opentelemetry:
-  otlp:
-    tracing:
-      exporter:
-        grpc:
-          endpoint: http://example.org
-          protocol: Grpc
-          timeout: 42
