+++ apollo-router/src/configuration/testdata/config_opentelemetry_otlp_tracing_grpc_full.yml
@@ -3,39 +3,3 @@ server:
 subgraphs:
   example:
     routing_url: http://foo
-opentelemetry:
-  otlp:
-    tracing:
-      exporter:
-        grpc:
-          endpoint: http://example.org
-          protocol: Grpc
-          timeout: 42
-          metadata:
-            - foo: bar
-            - foo: baz
-            - bar: baz
-      trace_config:
-        sampler:
-          ParentBased:
-            TraceIdRatioBased: 0.42
-        max_events_per_span: 1
-        max_attributes_per_span: 2
-        max_links_per_span: 3
-        max_attributes_per_event: 4
-        max_attributes_per_link: 5
-        resource:
-          attrs:
-            key1:
-              String: value
-            key2:
-              Bool: true
-            key3:
-              I64: 42
-            key4:
-              F64: 42.0
-            key5:
-              Array:
-                String:
-                  - value1
-                  - value2
