+++ apollo-router/src/configuration/testdata/config_opentelemetry_otlp_tracing_grpc_tls.yml
@@ -3,15 +3,3 @@ server:
 subgraphs:
   example:
     routing_url: http://foo
-opentelemetry:
-  otlp:
-    tracing:
-      exporter:
-        grpc:
-          tls_config:
-            ca:
-              env: SECRET_STUFF
-            cert:
-              file: /foo
-            key:
-              file: /bar
