+++ apollo-router/src/configuration/testdata/config_opentelemetry_otlp_tracing_http_basic.yml
@@ -3,8 +3,3 @@ server:
 subgraphs:
   example:
     routing_url: http://foo
-opentelemetry:
-  otlp:
-    tracing:
-      exporter:
-        http:
