+++ apollo-router/src/configuration/testdata/config_opentelemetry_otlp_tracing_http_common.yml
@@ -3,11 +3,3 @@ server:
 subgraphs:
   example:
     routing_url: http://foo
-opentelemetry:
-  otlp:
-    tracing:
-      exporter:
-        http:
-          endpoint: http://example.org
-          protocol: Grpc
-          timeout: 42
