+++ apollo-router/src/state_machine.rs
@@ -4,13 +4,11 @@ use super::state_machine::PrivateState::{Errored, Running, Startup, Stopped};
 use super::Event::{UpdateConfiguration, UpdateSchema};
 use super::FederatedServerError::{NoConfiguration, NoSchema};
 use super::{Event, FederatedServerError, State};
-use crate::configuration::{Configuration, SpaceportConfig};
+use crate::configuration::Configuration;
 use apollo_router_core::prelude::*;
 use apollo_router_core::Schema;
-use apollo_spaceport::server::ReportSpaceport;
 use futures::channel::mpsc;
 use futures::prelude::*;
-use std::pin::Pin;
 use std::sync::Arc;
 use Event::{NoMoreConfiguration, NoMoreSchema, Shutdown};
 
@@ -68,37 +66,6 @@ impl<RS> From<&PrivateState<RS>> for State {
     }
 }
 
-// For use when we have an external collector. Makes selecting over
-// events simpler
-async fn do_nothing(_addr_str: String) -> bool {
-    loop {
-        tokio::time::sleep(tokio::time::Duration::from_secs(3600)).await;
-    }
-    #[allow(unreachable_code)]
-    false
-}
-
-// For use when we have an internal collector.
-async fn do_listen(addr_str: String) -> bool {
-    tracing::info!("spawning an internal spaceport");
-    // Spawn a spaceport server to handle statistics
-    let addr = match addr_str.parse() {
-        Ok(a) => a,
-        Err(e) => {
-            tracing::error!("could not parse spaceport address: {}", e);
-            return false;
-        }
-    };
-
-    let spaceport = ReportSpaceport::new(addr);
-
-    if let Err(e) = spaceport.serve().await {
-        tracing::error!("spaceport did not terminate normally: {}", e);
-        return false;
-    }
-    true
-}
-
 impl<S, FA> StateMachine<S, FA>
 where
     S: HttpServerFactory,
@@ -121,49 +88,6 @@ where
         mut messages: impl Stream<Item = Event> + Unpin,
     ) -> Result<(), FederatedServerError> {
         tracing::debug!("Starting");
-        // Studio Agent Spaceport listener
-        // TODO: This code will be moving when we add support for Tower
-        // to the router. Leaving here until that time.
-        let (tx, mut rx) = tokio::sync::mpsc::channel::<SpaceportConfig>(1);
-
-        tokio::spawn(async move {
-            let mut current_listener = "".to_string();
-            let mut current_operation: fn(
-                msg: String,
-            )
-                -> Pin<Box<dyn Future<Output = bool> + Send>> = |msg| Box::pin(do_nothing(msg));
-
-            loop {
-                tokio::select! {
-                    biased;
-                    mopt = rx.recv() => {
-                        match mopt {
-                            Some(msg) => {
-                                tracing::info!(?msg);
-                                // Save our target listener for later use
-                                current_listener = msg.listener.clone();
-                                // Configure which function to call
-                                if msg.external {
-                                    current_operation = |msg| Box::pin(do_nothing(msg));
-                                } else {
-                                    current_operation = |msg| Box::pin(do_listen(msg));
-                                }
-                            },
-                            None => break
-                        }
-                    },
-                    x = current_operation(current_listener.clone()) => {
-                        // The only time a current_operation will return is failure
-                        // from the server. If this happens, wait for a while
-                        // then try again.
-                        tracing::info!(%x, "current_operation");
-                        tokio::time::sleep(tokio::time::Duration::from_secs(5)).await;
-                    }
-                };
-            }
-            tracing::info!("terminating spaceport loop");
-        });
-
         let mut state = Startup {
             configuration: None,
             schema: None,
@@ -184,29 +108,13 @@ where
                     .into_ok_or_err2(),
 
                 // Startup: Handle schema updates, maybe transition to running.
-                (Startup { schema, .. }, UpdateConfiguration(new_configuration)) => {
-                    // Only check for notify if we have graph configuration
-                    if new_configuration.graph.is_some() {
-                        match &new_configuration.spaceport {
-                            Some(v) => {
-                                tx.send(v.clone())
-                                    .await
-                                    .map_err(FederatedServerError::ServerSpaceportError)?;
-                            }
-                            None => {
-                                tx.send(Default::default())
-                                    .await
-                                    .map_err(FederatedServerError::ServerSpaceportError)?;
-                            }
-                        }
-                    }
-                    self.maybe_transition_to_running(Startup {
+                (Startup { schema, .. }, UpdateConfiguration(new_configuration)) => self
+                    .maybe_transition_to_running(Startup {
                         configuration: Some(*new_configuration),
                         schema,
                     })
                     .await
-                    .into_ok_or_err2()
-                }
+                    .into_ok_or_err2(),
 
                 // Startup: Missing configuration.
                 (
@@ -266,22 +174,6 @@ where
                     UpdateConfiguration(new_configuration),
                 ) => {
                     tracing::info!("Reloading configuration");
-                    // Only check for notify if we have graph configuration
-                    if new_configuration.graph.is_some() {
-                        match &new_configuration.spaceport {
-                            Some(v) => {
-                                tx.send(v.clone())
-                                    .await
-                                    .map_err(FederatedServerError::ServerSpaceportError)?;
-                            }
-                            None => {
-                                tx.send(Default::default())
-                                    .await
-                                    .map_err(FederatedServerError::ServerSpaceportError)?;
-                            }
-                        }
-                    }
-
                     self.reload_server(
                         configuration,
                         schema,
