+++ apollo-router/tests/fixtures/supergraph_config.yaml
@@ -15,5 +15,3 @@ subgraphs:
     routing_url: http://localhost:4004/graphql
     schema:
       file: ./inventory.graphql
-opentelemetry:
-  jaeger:
