+++ apollo-router/tests/integration_tests.rs
@@ -1,4 +1,5 @@
 use apollo_router::configuration::Configuration;
+use apollo_router::get_dispatcher;
 use apollo_router::reqwest_subgraph_service::ReqwestSubgraphService;
 use apollo_router_core::prelude::*;
 use apollo_router_core::{
@@ -308,13 +309,7 @@ async fn query_rust(
         serde_yaml::from_str::<Configuration>(include_str!("fixtures/supergraph_config.yaml"))
             .unwrap();
     let counting_registry = CountingServiceRegistry::new();
-    let dispatcher = config
-        .subscriber
-        .clone()
-        .map(tracing::Dispatch::new)
-        .unwrap_or_default();
-
-    let mut builder = PluggableRouterServiceBuilder::new(schema, 10, dispatcher);
+    let mut builder = PluggableRouterServiceBuilder::new(schema, 10);
     for (name, subgraph) in &config.subgraphs {
         let cloned_counter = counting_registry.clone();
         let cloned_name = name.clone();
@@ -329,6 +324,7 @@ async fn query_rust(
         builder = builder.with_subgraph_service(name, service);
     }
 
+    builder = builder.with_dispatcher(get_dispatcher());
     let (mut router, _) = builder.build().await;
 
     let stream = router.ready().await.unwrap().call(request).await.unwrap();
